Rename files as snake_case according to google style guide
