+++ results/hojongs/src/service/url_param_service.ts
@@ -1,5 +1,5 @@
 export default class UrlParamService {
-    getAccessTokenOrThrow() {
+    getAccessTokenOrThrow(): string {
         const accessToken = new URL(location.href).searchParams.get("accessToken");
         if (accessToken == null) throw "accessToken was null";
         return accessToken;
